# MongoDB_API/MongoDB_API/Models/Product.cs
using Microsoft.AspNetCore.Mvc;
using MongoDB.Bson;
using MongoDB.Bson.Serialization.Attributes;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace MongoDB_API.Models
{
    public class Product
    {
        [BsonId]
        //[BsonRepresentation(MongoDB.Bson.BsonType.ObjectId)]

        [FromQuery] 
        public ObjectId Id { get; set; }

        public string Name { get; set; }
        public int Price { get; set; }

        public int Cantidad { get; set; }

        public string Category { get; set; }

        
    }
}